Vuln scanners, split up UI files, scheduler popup
Added a new combo box to the main window that allows the user to select
from a list of vulnerability scanners to run on the target host. The
combobox is populated from a list of available scanners

Also split up the UI files into separate files for each dialog and used
PyQt6.uic.loadUi() to load them directly. This eliminates the need to
convert them to .py files first and should make it easier to maintain
the UI code going forward.

Finally, implemented the basic structure for the scheduler popup dialog,
although it is currently non-functional.